@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from "react-router-dom";
 import { Authenticator } from '@aws-amplify/ui-react';
 import App from './App';
@@ -7,13 +7,13 @@ import "@fontsource/karla/400.css";
 import "@fontsource/karla/700.css";
 import './index.css';
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
   <React.StrictMode>
     <BrowserRouter>
       <Authenticator.Provider>
         <App />
       </Authenticator.Provider>
     </BrowserRouter>
-  </React.StrictMode>,
-  document.getElementById('root')
-)
+  </React.StrictMode>
+);
