@@ -1,7 +1,7 @@
 import React from 'react';
 import { Outlet } from "react-router-dom";
-import Header from '../components/Header';
-import Footer from '../components/Footer';
+import Header from './Header';
+import Footer from './Footer';
 import '@aws-amplify/ui-react/styles.css'; // this adds styles for the amplify UI
 
 function Layout() {
